test: script to track performance regression
Adapted from pyrealm. Allows to compare two profiled runs of the code
and highlights which functions performance dropped below a threshold.

Note that at the memonet DEMENTpy exhibits high variance in runtime
of individual functions (observed up to 15%) so the check with default
settings (5% tolerance) will likley fail for the same code versions.