feat: change output format to CSV
Closes #35

Writes a set of CSV files to a folder instead of 'pickle'ing the output
data class. Is more human-readable and should make reading data by
post-processing scripts easier.

The drawback is that if new members of unsupported type are introduced
they may not be written (a warning is raised in that case) and adding
extra case to the export functions might be necessary.